Changed: Separate out the app so it can be reused between webpack + next
