+++ src/index.jsx
@@ -1,33 +1,14 @@
 // @flow
-import React from 'react';
 import { render } from 'react-dom';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import { combineReducers, createStore } from 'redux';
-import { Provider } from 'react-redux';
+import base from './base';
 
 import '@babel/polyfill';
-
-import App from './App/AppContainer';
-import { appReducer } from './App/AppReducer';
-
-const reducers = combineReducers({
-  app: appReducer,
-});
-const store = createStore(reducers);
-
 
 const profileAppElement = document.getElementById('profile-app');
 
 if (profileAppElement instanceof Element) {
   render(
-    <Provider store={store}>
-      <BrowserRouter>
-        <Switch>
-          <Route path="/:name" component={App} />
-          <Route path="/" component={App} />
-        </Switch>
-      </BrowserRouter>
-    </Provider>,
+    base(),
     profileAppElement,
   );
 }
